Rename event and move into separated module
