+++ packages/twenty-server/src/engine/core-modules/workflow/workflow-trigger.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { WorkflowTriggerResolver } from 'src/engine/core-modules/workflow/workflow-trigger.resolver';
-import { WorkflowTriggerService } from 'src/engine/core-modules/workflow/workflow-trigger.service';
+import { WorkflowTriggerService } from 'src/modules/workflow/standard-objects/workflow-trigger/workflow-trigger.service';
 
 @Module({
   providers: [WorkflowTriggerService, WorkflowTriggerResolver],
