+++ packages/twenty-server/src/engine/core-modules/workflow/workflow-trigger.resolver.ts
@@ -2,10 +2,10 @@ import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
 
 import { workflowTriggerGraphqlApiExceptionHandler } from 'src/engine/core-modules/workflow/utils/workflow-trigger-graphql-api-exception-handler.util';
-import { WorkflowTriggerService } from 'src/engine/core-modules/workflow/workflow-trigger.service';
 import { Workspace } from 'src/engine/core-modules/workspace/workspace.entity';
 import { AuthWorkspace } from 'src/engine/decorators/auth/auth-workspace.decorator';
 import { JwtAuthGuard } from 'src/engine/guards/jwt.auth.guard';
+import { WorkflowTriggerService } from 'src/modules/workflow/standard-objects/workflow-trigger/workflow-trigger.service';
 
 @UseGuards(JwtAuthGuard)
 @Resolver()
