+++ packages/twenty-server/src/modules/workflow/standard-objects/workflow-trigger/workflow-trigger.exception.ts
